Split sv_phys.c, sv_move.c and world.c into topic modules (core/client/ent/push, monster_move/ai, area/hull/move)

+++ src/engine/Server_side/server_priv.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-    void SV_StartParticle(vec3_t org, vec3_t dir, int color, size_t count);
+    void SV_StartParticle(vec3_t org, vec3_t dir, uint8_t color, uint8_t count);
     void SV_SetIdealPitch();
     void SV_MoveToGoal();
     bool SV_movestep(edict_p ent, vec3_t move, bool relink);
